@@ -45,9 +45,9 @@
 /***************************************************************************************************
  * Zone JS is required by default for Angular itself.
  */
-import 'zone.js';  // Included with Angular CLI.
+import 'zone.js'; // Included with Angular CLI.
 
-
+import '@angular/localize/init';
 /***************************************************************************************************
  * APPLICATION IMPORTS
  */
